+++ packages/types/src/json.ts
@@ -26,8 +26,8 @@ import type { PhoneCodeChannel } from './phoneCodeChannel';
 import type { SamlIdpSlug } from './saml';
 import type { SessionStatus, SessionTask } from './session';
 import type { SessionVerificationLevel, SessionVerificationStatus } from './sessionVerification';
-import type { SignInFirstFactor, SignInJSON, SignInSecondFactor } from './signIn';
-import type { SignUpField, SignUpIdentificationField, SignUpStatus } from './signUp';
+import type { SignInJSON } from './signIn';
+import type { SignInFirstFactor, SignInSecondFactor } from './signInCommon';
 import type { BoxShadow, Color, EmUnit, FontWeight, HexColor } from './theme';
 import type { UserSettingsJSON } from './userSettings';
 import type { CamelToSnake } from './utils';
