+++ packages/types/src/signInCommon.ts
@@ -1,0 +1,193 @@
+import type {
+  BackupCodeAttempt,
+  BackupCodeFactor,
+  EmailCodeAttempt,
+  EmailCodeConfig,
+  EmailCodeFactor,
+  EmailLinkConfig,
+  EmailLinkFactor,
+  EnterpriseSSOConfig,
+  EnterpriseSSOFactor,
+  OAuthConfig,
+  OauthFactor,
+  PasskeyAttempt,
+  PassKeyConfig,
+  PasskeyFactor,
+  PasswordAttempt,
+  PasswordFactor,
+  PhoneCodeAttempt,
+  PhoneCodeConfig,
+  PhoneCodeFactor,
+  PhoneCodeSecondFactorConfig,
+  ResetPasswordEmailCodeAttempt,
+  ResetPasswordEmailCodeFactor,
+  ResetPasswordEmailCodeFactorConfig,
+  ResetPasswordPhoneCodeAttempt,
+  ResetPasswordPhoneCodeFactor,
+  ResetPasswordPhoneCodeFactorConfig,
+  SamlConfig,
+  SamlFactor,
+  TOTPAttempt,
+  TOTPFactor,
+  Web3Attempt,
+  Web3SignatureConfig,
+  Web3SignatureFactor,
+} from './factors';
+import type {
+  EmailAddressIdentifier,
+  PhoneNumberIdentifier,
+  UsernameIdentifier,
+  Web3WalletIdentifier,
+} from './identifiers';
+import type {
+  BackupCodeStrategy,
+  EmailCodeStrategy,
+  EmailLinkStrategy,
+  EnterpriseSSOStrategy,
+  GoogleOneTapStrategy,
+  OAuthStrategy,
+  PasskeyStrategy,
+  PasswordStrategy,
+  PhoneCodeStrategy,
+  ResetPasswordEmailCodeStrategy,
+  ResetPasswordPhoneCodeStrategy,
+  SamlStrategy,
+  TicketStrategy,
+  TOTPStrategy,
+  Web3Strategy,
+} from './strategies';
+import type { StartEmailLinkFlowParams } from './verification';
+
+export type SignInStatus =
+  | 'needs_identifier'
+  | 'needs_first_factor'
+  | 'needs_second_factor'
+  | 'needs_new_password'
+  | 'complete';
+
+export type SignInIdentifier =
+  | UsernameIdentifier
+  | EmailAddressIdentifier
+  | PhoneNumberIdentifier
+  | Web3WalletIdentifier;
+
+export type SignInFirstFactor =
+  | EmailCodeFactor
+  | EmailLinkFactor
+  | PhoneCodeFactor
+  | PasswordFactor
+  | PasskeyFactor
+  | ResetPasswordPhoneCodeFactor
+  | ResetPasswordEmailCodeFactor
+  | Web3SignatureFactor
+  | OauthFactor
+  | SamlFactor
+  | EnterpriseSSOFactor;
+
+export type SignInSecondFactor = PhoneCodeFactor | TOTPFactor | BackupCodeFactor;
+
+export interface UserData {
+  firstName?: string;
+  lastName?: string;
+  imageUrl?: string;
+  hasImage?: boolean;
+}
+
+export type SignInFactor = SignInFirstFactor | SignInSecondFactor;
+
+export type PrepareFirstFactorParams =
+  | EmailCodeConfig
+  | EmailLinkConfig
+  | PhoneCodeConfig
+  | Web3SignatureConfig
+  | PassKeyConfig
+  | ResetPasswordPhoneCodeFactorConfig
+  | ResetPasswordEmailCodeFactorConfig
+  | OAuthConfig
+  | SamlConfig
+  | EnterpriseSSOConfig;
+
+export type AttemptFirstFactorParams =
+  | PasskeyAttempt
+  | EmailCodeAttempt
+  | PhoneCodeAttempt
+  | PasswordAttempt
+  | Web3Attempt
+  | ResetPasswordPhoneCodeAttempt
+  | ResetPasswordEmailCodeAttempt;
+
+export type PrepareSecondFactorParams = PhoneCodeSecondFactorConfig;
+
+export type AttemptSecondFactorParams = PhoneCodeAttempt | TOTPAttempt | BackupCodeAttempt;
+
+export type SignInCreateParams = (
+  | {
+      strategy: OAuthStrategy | SamlStrategy | EnterpriseSSOStrategy;
+      redirectUrl: string;
+      actionCompleteRedirectUrl?: string;
+      identifier?: string;
+      oidcPrompt?: string;
+      oidcLoginHint?: string;
+    }
+  | {
+      strategy: TicketStrategy;
+      ticket: string;
+    }
+  | {
+      strategy: GoogleOneTapStrategy;
+      token: string;
+    }
+  | {
+      strategy: PasswordStrategy;
+      password: string;
+      identifier: string;
+    }
+  | { strategy: PasskeyStrategy }
+  | {
+      strategy:
+        | PhoneCodeStrategy
+        | EmailCodeStrategy
+        | Web3Strategy
+        | ResetPasswordEmailCodeStrategy
+        | ResetPasswordPhoneCodeStrategy;
+      identifier: string;
+    }
+  | {
+      strategy: EmailLinkStrategy;
+      identifier: string;
+      redirectUrl?: string;
+    }
+  | {
+      identifier: string;
+    }
+  | { transfer?: boolean }
+) & { transfer?: boolean };
+
+export type ResetPasswordParams = {
+  password: string;
+  signOutOfOtherSessions?: boolean;
+};
+
+export type AuthenticateWithPasskeyParams = {
+  flow?: 'autofill' | 'discoverable';
+};
+
+export interface SignInStartEmailLinkFlowParams extends StartEmailLinkFlowParams {
+  emailAddressId: string;
+}
+
+export type SignInStrategy =
+  | PasskeyStrategy
+  | PasswordStrategy
+  | ResetPasswordPhoneCodeStrategy
+  | ResetPasswordEmailCodeStrategy
+  | PhoneCodeStrategy
+  | EmailCodeStrategy
+  | EmailLinkStrategy
+  | TicketStrategy
+  | Web3Strategy
+  | TOTPStrategy
+  | BackupCodeStrategy
+  | OAuthStrategy
+  | SamlStrategy
+  | EnterpriseSSOStrategy;
